requirements.txt
tensorflow==1.3.0
click==6.7
filelock==2.0.11
Flask==0.12.2
gunicorn==19.7.1
h5py==2.7.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
nose==1.3.7
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
protobuf==3.4.0
six==1.10.0
Werkzeug==0.12.2
